[Fix] 로그인한 계정 조회 기능
- LoginAccountId
로그인한 계정의 정보를 가져오는 애너테이션 구현

- Controller
기존 URL로 memberId 를 받아오던 방식을 LoginAccountId 애너테이션을 이용하도록 변경

+++ server/src/main/java/com/seb40pre023/domain/answer/service/AnswerService.java
@@ -33,8 +33,13 @@ public class AnswerService {
     }
 
     // 회원이 질문글에 작성한 답변 수정
-    public Answer updateAnswer(Answer answer) {
+    public Answer updateAnswer(Answer answer, Long memberId) {
         Answer findAnswer = findVerifiedAnswer(answer.getAnswerId());
+
+        if (findAnswer.getMember().getMemberId() != memberId){
+            throw new BusinessLogicException(ExceptionCode.MEMBER_NOT_FOUND);
+        }
+
         Optional.ofNullable(answer.getContent())
                 .ifPresent(content -> findAnswer.setContent(content));
 
@@ -54,7 +59,13 @@ public class AnswerService {
         return answers;
     }
 
-    public void deleteAnswer(Long answerId) {
+    public void deleteAnswer(Long answerId, Long memberId) {
+
+        Answer findAnswer = findVerifiedAnswer(answerId);
+
+        if (findAnswer.getMember().getMemberId() != memberId){
+            throw new BusinessLogicException(ExceptionCode.MEMBER_NOT_FOUND);
+        }
 
         answerRepository.deleteById(answerId);
     }
